refactor(daemon): give the JSON-Schema tool vocabulary a neutral home
`src/mcp/tool-descriptor.ts` was the last edge closing the memory -> mcp
import cycle: `memory/tools.ts` and every memory provider reached into
`mcp/` purely for the MCP-agnostic JSON-Schema vocabulary.

Move it to `src/tool-schema/descriptor.ts`, a dependency-free leaf, and
retarget all 15 importers. `JSONValue` is now defined locally instead of
imported from `@modelcontextprotocol/server`, so the module has zero
imports. No re-export shell is left behind in `mcp/`.
