+++ packages/daemon/src/memory/providers/dispatching.ts
@@ -1,6 +1,6 @@
 import type { AgentMemoryBinding, CaptureReceipt, ExternalMemoryBinding, MemoryEntry } from '@agentconnect.md/protocol'
 import type { RuntimeDef } from '../../config/config-schema.js'
-import type { ToolDescriptor } from '../../mcp/tool-descriptor.js'
+import type { ToolDescriptor } from '../../tool-schema/descriptor.js'
 import { MEMORY_TOOLS } from '../tools.js'
 import type { MemoryFs, MemoryWriteSource } from '../store.js'
 import {
