+++ packages/daemon/src/memory/providers/none.ts
@@ -1,6 +1,6 @@
 import type { MemoryEntry } from '@agentconnect.md/protocol'
 import type { RuntimeDef } from '../../config/config-schema.js'
-import type { ToolDescriptor } from '../../mcp/tool-descriptor.js'
+import type { ToolDescriptor } from '../../tool-schema/descriptor.js'
 import {
   MemoryProviderUnavailableError,
   type MemoryProvider,
