+++ packages/daemon/src/tool-schema/descriptor.ts
@@ -1,4 +1,5 @@
-import type { JSONValue } from '@modelcontextprotocol/server'
+/** A JSON value, defined locally so this vocabulary stays a dependency-free leaf. */
+export type JsonValue = string | number | boolean | null | JsonValue[] | { [key: string]: JsonValue }
 
 /** A tool descriptor in MCP's `tools/list` shape: name, model-facing description, and an argument JSON Schema. */
 export interface ToolDescriptor {
@@ -7,9 +8,9 @@ export interface ToolDescriptor {
   inputSchema: ObjectToolSchema | ObjectUnionSchema
 }
 
-export type ToolProperties = Record<string, JSONValue>
+export type ToolProperties = Record<string, JsonValue>
 
-export interface ObjectToolSchema extends Record<string, JSONValue> {
+export interface ObjectToolSchema extends Record<string, JsonValue> {
   type: 'object'
   properties: ToolProperties
   required: string[]
@@ -17,7 +18,7 @@ export interface ObjectToolSchema extends Record<string, JSONValue> {
 }
 
 /** A root object schema whose `oneOf` branches are mutually exclusive call modes (used by sendMessage). */
-export interface ObjectUnionSchema extends Record<string, JSONValue> {
+export interface ObjectUnionSchema extends Record<string, JsonValue> {
   type: 'object'
   oneOf: ObjectToolSchema[]
 }
